Fixes after review

+++ scrapper/src/main/java/edu/java/services/jdbc/JdbcLinkService.java
@@ -1,5 +1,6 @@
 package edu.java.services.jdbc;
 
+import edu.java.api.exceptions.DoubleLinkException;
 import edu.java.api.exceptions.NoSuchLinkException;
 import edu.java.model.Link;
 import edu.java.repository.jdbc.JdbcLinkRepository;
@@ -7,12 +8,10 @@ import edu.java.services.LinkService;
 import java.time.OffsetDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Slf4j
 @RequiredArgsConstructor
 public class JdbcLinkService implements LinkService {
     private final JdbcLinkRepository jdbcLinkRepository;
@@ -26,7 +25,9 @@ public class JdbcLinkService implements LinkService {
         newLink.setUrl(url);
 
         Link link = jdbcLinkRepository.findByUrl(url).orElseGet(() -> jdbcLinkRepository.addLink(newLink));
-        jdbcLinkRepository.addLinkToChat(tgChatId, link.getId());
+        if (!jdbcLinkRepository.addLinkToChat(tgChatId, link.getId())) {
+            throw new DoubleLinkException("Link is already being tracked");
+        }
 
         return link;
     }
@@ -35,7 +36,9 @@ public class JdbcLinkService implements LinkService {
     @Transactional
     public Link remove(long tgChatId, String url) {
         Link link = jdbcLinkRepository.findByUrl(url).orElseThrow(() -> new NoSuchLinkException("Link was not found"));
-        jdbcLinkRepository.removeLinkByChat(tgChatId, link.getId());
+        if (!jdbcLinkRepository.removeLinkByChat(tgChatId, link.getId())) {
+            throw new NoSuchLinkException("No such link");
+        }
 
         if (!jdbcLinkRepository.findLinkInAllChats(link.getId())) {
             jdbcLinkRepository.removeLink(link.getId());
@@ -46,6 +49,6 @@ public class JdbcLinkService implements LinkService {
 
     @Override
     public List<Link> listAll(long tgChatId) {
-        return jdbcLinkRepository.findAllByChat(tgChatId);
+        return jdbcLinkRepository.findLinksByChatsId(tgChatId);
     }
 }
